separates classes into separate files
