+++ torch/CMakeLists.txt
@@ -158,6 +158,8 @@ set(TORCH_SRCS
   ${TORCH_SRC_DIR}/csrc/jit/fusers/cpu/cpu_fuser.cpp
   ${TORCH_SRC_DIR}/csrc/jit/fusers/cuda/cuda_fusion_pass.cpp
   ${TORCH_SRC_DIR}/csrc/jit/fusers/cuda/cuda_fuser.cpp
+  ${TORCH_SRC_DIR}/csrc/jit/fusers/cuda/tensor_desc.cpp
+  ${TORCH_SRC_DIR}/csrc/jit/fusers/cuda/cuda_fusion_function.cpp
   ${TORCH_SRC_DIR}/csrc/jit/passes/inplace_check.cpp
   ${TORCH_SRC_DIR}/csrc/jit/passes/loop_unrolling.cpp
   ${TORCH_SRC_DIR}/csrc/jit/passes/lower_grad_of.cpp
